fix: split transport overlay paint under the 300-line flame file ceiling
Keep resource and improvement overlay paint in overlays.dart so the Quality flame size ratchet can pass.
