+++ app/lib/features/game/flame/region_map/region_map_component_render_core_overlays.dart
@@ -1,15 +1,13 @@
 import 'dart:math' as math;
-import 'package:colonizethis_data/colonizethis_data.dart';
 
 import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
 import 'package:colonizethis_map/colonizethis_map.dart';
-import 'package:flutter/material.dart';
-import '../caches/resource_icon_cache.dart';
-import '../tilesets/tilesets.dart';
-import 'region_map_component.dart';
-import 'region_map_component_render_core.dart';
 import 'package:colonizethis_world/colonizethis_world.dart'
     show resourceIdVisibleInPlayerView;
+import 'package:flutter/material.dart';
+
+import '../caches/resource_icon_cache.dart';
+import 'region_map_component.dart';
 
 /// Paint colour for an improvement corner mark (Refs #4408).
 Color improvementMarkPaintColor({
@@ -25,163 +23,7 @@ Color improvementMarkPaintColor({
   return Colors.black;
 }
 
-extension CtRegionMapRenderCoreTransportFeature on CtRegionMapComponent {
-  void paintL1PlainsInteriorResourceVariantOverlays(Canvas canvas) {
-    for (final cell in region.cells) {
-      if (cell.isSea) continue;
-      if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-          regionMapComponentVisibilityForTerrain(this, cell) ==
-              TileVisibility.unrevealed) {
-        continue;
-      }
-      if (cell.terrainType != TerrainType.plains) continue;
-      if (isPlainTerrainAtDesertTransitionWangCell(cell)) continue;
-
-      final plainsVariantKey = landInteriorPlainsVariantTileKey(cell);
-      if (plainsVariantKey == null) continue;
-
-      final standaloneTile = terrainTilesetCache.getStandaloneTileByKey(
-        plainsVariantKey,
-      );
-      if (standaloneTile == null) {
-        throw StateError(
-          'Missing required plains terrain variant tile: $plainsVariantKey',
-        );
-      }
-      final left = cell.x * cellSize;
-      final top = cell.y * cellSize;
-      final dstRect = Rect.fromLTWH(left, top, cellSize, cellSize);
-      final overlayPaint = Paint();
-      if (shouldApplyFogToInteriorPlainsVariantOverlay(
-        visibilityMode: visibilityMode,
-        tileVisibility: regionMapComponentVisibilityForTerrain(this, cell),
-      )) {
-        overlayPaint.colorFilter = ColorFilter.mode(
-          Color.fromRGBO(0, 0, 0, RegionMapPalette.fogOverlayOpacity),
-          BlendMode.darken,
-        );
-      }
-      final srcRect = Rect.fromLTWH(
-        0,
-        0,
-        standaloneTile.image.width.toDouble(),
-        standaloneTile.image.height.toDouble(),
-      );
-      canvas.drawImageRect(
-        standaloneTile.image,
-        srcRect,
-        dstRect,
-        overlayPaint,
-      );
-    }
-  }
-
-  bool isPlainTerrainAtDesertTransitionWangCell(CellViewData cell) {
-    final nearDesertCorner = getCornerValues(
-      cell.x,
-      cell.y,
-      (c) => !c.isSea && c.terrainType == TerrainType.desert,
-    );
-    return !nearDesertCorner.same && nearDesertCorner.value;
-  }
-
-  void paintTransportOverlayTiles(Canvas canvas) {
-    if (!shouldRenderTransportOverlay(flags: mapBaseLayerFlags) ||
-        !transportOverlayTilesetCache.isLoaded) {
-      return;
-    }
-    for (final cell in region.cells) {
-      final tileVisibility = regionMapComponentVisibilityForTerrain(this, cell);
-      if (!shouldPaintTransportOverlayForCell(
-        cell: cell,
-        visibilityMode: visibilityMode,
-        tileVisibility: tileVisibility,
-      )) {
-        continue;
-      }
-      final roadLevel = cell.roadLevel ?? 0;
-      final family = isRailTransportLevel(roadLevel)
-          ? TransportTileFamily.rail
-          : TransportTileFamily.road;
-      final tileset = transportOverlayTilesetCache.getTileset(family);
-      if (tileset == null) {
-        continue;
-      }
-      final mask = computeTransportConnectivityMask(
-        x: cell.x,
-        y: cell.y,
-        getCellAt: getCellAt,
-      );
-      final srcRect = tileset.tileRectForMask(mask);
-      if (srcRect == null) {
-        regionMapComponentLifecycleLog.w(
-          'Transport tile missing for family=$family mask=$mask',
-        );
-        continue;
-      }
-      final tileLeft = cell.x * cellSize;
-      final tileTop = cell.y * cellSize;
-      final dstRect = Rect.fromLTWH(tileLeft, tileTop, cellSize, cellSize);
-      canvas.drawImageRect(
-        tileset.image,
-        srcRect,
-        dstRect,
-        resourceOverlayPaintForCell(cell),
-      );
-    }
-  }
-
-  void paintFeatureCell(Canvas canvas, CellViewData cell) {
-    final left = cell.x * cellSize;
-    final top = cell.y * cellSize;
-
-    if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-        regionMapComponentVisibilityForTerrain(this, cell) ==
-            TileVisibility.unrevealed) {
-      return;
-    }
-
-    final terrain = cell.terrainType;
-    if (terrain == null || !regionMapComponentIsFeatureTerrain(terrain)) return;
-
-    final overlayTileKey = featureOverlayTileKey(
-      terrain: terrain,
-      resourceId: cell.resourceId,
-      improvementLevel: cell.improvementLevel,
-    );
-    final standaloneTile =
-        terrainTilesetCache.getStandaloneTileByKey(overlayTileKey) ??
-        terrainTilesetCache.getStandaloneTile(terrain);
-    if (standaloneTile == null) {
-      regionMapComponentLifecycleLog.w(
-        'Feature overlay tile missing for key=$overlayTileKey terrain=$terrain; '
-        'skipping feature overlay for cell (${cell.x}, ${cell.y})',
-      );
-      return;
-    }
-
-    final paint = Paint();
-    if (shouldApplyFogToFeatureOverlay(
-      visibilityMode: visibilityMode,
-      tileVisibility: regionMapComponentVisibilityForTerrain(this, cell),
-      terrain: terrain,
-    )) {
-      paint.colorFilter = ColorFilter.mode(
-        Color.fromRGBO(0, 0, 0, RegionMapPalette.fogOverlayOpacity),
-        BlendMode.darken,
-      );
-    }
-
-    final srcRect = Rect.fromLTWH(
-      0,
-      0,
-      standaloneTile.image.width.toDouble(),
-      standaloneTile.image.height.toDouble(),
-    );
-    final dstRect = Rect.fromLTWH(left, top, cellSize, cellSize);
-    canvas.drawImageRect(standaloneTile.image, srcRect, dstRect, paint);
-  }
-
+extension CtRegionMapRenderCoreOverlays on CtRegionMapComponent {
   Paint resourceOverlayPaintForCell(CellViewData cell) {
     final paint = Paint();
     if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
@@ -196,9 +38,7 @@ extension CtRegionMapRenderCoreTransportFeature on CtRegionMapComponent {
     }
     return paint;
   }
-}
 
-extension CtRegionMapRenderCoreOverlays on CtRegionMapComponent {
   String? _resourceIdForMapIcon(CellViewData cell) {
     final raw = cell.resourceId;
     if (raw == null) return null;
